Add syndication workflow (#38)

+++ _netlify/plugins/add-dot-env/index.js
@@ -2,6 +2,12 @@ const fs = require('fs')
 
 module.exports = {
   onPreBuild: () => {
-    fs.writeFileSync('.env', '', { encoding: 'utf-8' })
+    try {
+      if(!fs.existsSync('.env')) {
+        fs.writeFileSync('.env', '', { encoding: 'utf-8' })
+      }
+    } catch (err) {
+      console.error(err);
+    }
   }
 }
